requirements.txt
streamlit==1.31.0
PyPDF2==3.0.1
pdf2image==1.16.3
paddleocr==2.7.0.3
sentence-transformers==2.2.2
scikit-learn==1.3.2
pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
pytesseract==0.3.10